requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery
redis
python-dotenv
httpx
